requirements.txt
#Python Generic
python-dotenv==1.0.0
chardet==5.2.0

#Open AI
openai==1.30.1

#Langchain
langchain==0.2.0
langchain-community==0.2.0
langchain-experimental==0.0.59
langchain-openai==0.1.7
langgraph==0.0.53

#Streamlit
streamlit==1.34.0

# azure
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-search-documents==11.4.0